/pages/routes/index.js now holding paths object, mapped navLinks, routes. Manualy update /routes/index instead of app.js to create new endpoint
